% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAPEabm-package.R
\docType{package}
\name{SHAPEabm-package}
\alias{SHAPEabm-package}
\alias{SHAPEabm}
\title{What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
Some description of the package
}
\examples{
## An example function:
df <- example_function()
df

## To cite this package in publications use:
citation("SHAPEabm")

}
\references{
Some article that you might want users to look at
}
\seealso{
\code{\link[neatpkg]{pkg_new}} for the function that created this package
}
\author{
\strong{Maintainer}: Joe Bloggs \email{joe@bloggs.com}

}
\keyword{internal}
